Import the DB dump using `app-run`
Alternative approach to import the DB, using the command
`tsuru app-run` to install the postgresql client, download
the dump and import it. In this case the import happens using the app user.

This approach has the benefit of not depending on SSH to the DB server.

In order to get the dump, it must be served via HTTP from a
given location, optionally with some authentication Header
like `Referer` or `Token`