Add task to set WEB_CONCURRENCY for postgresapi
The postgresapi uses a unique connection per thread to connect to
the database, which can lead to errors while creating/binding services
to the applications in parallel.

To workaround the issue, we increase the `$WEB_CONCURRENCY` variable
which will [increase the number of workers
gunicorn](http://gunicorn-docs.readthedocs.org/en/latest/settings.html#workers)
and reduce the chances of failure.